Fixes for issues detected by code scanning.

+++ src/jandy/messages/iaq/iaq_message_page_button.cpp
@@ -68,7 +68,7 @@ namespace AqualinkAutomate::Messages
 		{
 			LogDebug(Channel::Messages, "IAQMessage_PageButton is too short to deserialise context of ButtonName.");
 		}
-		else if ((JandyMessage::MINIMUM_PACKET_LENGTH + 1 + 1 + 1) > message_bytes.size())
+		else if (static_cast<uint64_t>(JandyMessage::MINIMUM_PACKET_LENGTH + 1 + 1 + 1) > message_bytes.size())
 		{
 			LogDebug(Channel::Messages, "IAQMessage_PageButton is too short to deserialise content of LineText");
 		}
